requirements.txt
Babel
Django
django-treebeard
djangorestframework
pytz
